Fixed an error parsing trailing key-value pairs.
This error would arise when a file was terminated by a key-value pair.
For example:
  contents = 'key=value'

The type of token received here is EOF, but we should still parse the
value before exiting.
Previously, we exited with an error status.

Also fixed an error where tokens associated with existing keys were not
correctly parsed before being added to the sequence value.