Changed parsing of repeated keys with dict values
This situation arises where an object in the save file has multiple
identical keys, and the corresponding values are themselves objects.

For example:
    {
        key={
            one=1
        }
        key={
            one=2
        }
    }

This is an artefact of an...odd choice of save format by Paradox.

Previously, the *later* of the two values would be taken.
In other words, the result from the above example would be the dict:
    {
        'key': {
            'one': 2,
        },
    }

This behaviour is now reversed; the earlier value is taken.

This is a quick fix for issue #3:
* when
  (a) a country changes tag; and
  (b) that country had previously captured territory;
* then
  * EU4 will 'revise history' at the time of the tag change
  * this fakes a change of ownership to the new tag at the date of capture

The faked ownership change always comes after the real one, so switching
this parsing behaviour acts as a hotfix for the issue.

In reality, proper support for repeated events should be added.